sql: Add privileges for system privileges (#20294)
This commit restricts granting and revoking system to superusers only.
To grant/revoke a privilege on some object, generally the
grantor/revoker must be the owner of that object (or have a grant option
on that object which isn't implemented in Materialize yet). There is no
owner of the entire system, so it's only reasonable to restrict
granting/revoking system privileges to superusers.