environmentd: improve long-lived connection logic for websockets
The adapter client handling was initially written only for single-shot
HTTP endpoints and so didn't do various things done in the pgwire
protocol. But since websockets exist we need this logic but never wrote
tests for it.

Teach websockets to correctly close implicit transactions and also reset
the client state at the beginning of each message.

Fixes #20262