stash-debug: use JSON
When actually using this tool, correctly specifying hex-encoded protobuf
is too hard to be useful. We could have taught the tool to print the
proto keys which could have been used for dump, but would not have been
helpful for edit setting a new value.

Use JSON instead for all interfaces. Teach dump to output both the rust
Debug of the original structs and also a oneliner that can be
copy-pasted into a terminal for use with edit and delete.