Document coordinator interface.
This documents some of the control flow and assupmtions between pgwire
and the coordinator that I personally found non-obvious, and so
hopefully it can help someone else.

The biggest deficiency of this text is that it does not discuss
transactions at all. I did not feel like my knowledge of those was
sufficient for documenting them properly, but it would be great if
someone else added that information in the future.