adapter: Validate deferred plan before sequencing
The adapter uses a global write lock to linearize read then write
statements. If a read then write statement fails to acquire the write
lock then it is pushed into a queue of deferred statements. Once the
statement has acquired the write lock, it is popped off the queue and
sequenced (i.e. executed).

While the deferred statement is on the queue, we allow DDL statements
to execute. After we pop the statement off the queue we were not
checking to make sure all the statements dependencies still existed.
If a dependency was deleted while the statement was in the queue, then
the statement would cause a panic during sequencing because it assumed
all of its dependencies were validated during planning.

This commit adds a step after the statement is popped off of the queue
to revalidate all dependencies and check that they still exist. If a
dependency was deleted while it was waiting, then an error is returned
to the user.

There are various points in the Coordinator where either a statement
or plan is taken off of the main Coordinator thread for some
asynchronous processing/waiting. Roughly each point has its own
bespoke code for re-validating the statement or plan. Issues like this
probably suggest that we should implement a more general framework for
taking statements and plans off of the main Coordinator thread. For now
another bespoke validation is probably good enough.

Fixes #20402