adapter: add builtin views required by PopSQL
PopSQL is a modern SQL notebook with both desktop and cloud editions.
This commit adds several information_schema and pg_catalog views that
PopSQL relies on.

With these additions, nearly all features of PopSQL work when using the
desktop edition. The cloud edition additionally requires that we support
the `iso_8601` IntervalStyle, which is proving tricky and will be
addressed in a separate commit.