Improve invalid accumulation error in DistinctBy
This commit improves the error reporting of invalid accumulations in
DistinctBy. Previously, the error message would output the empty
values associated with a key, since in the case of DistinctBy the
entire row is moved into the key. Now, the error message reported as
a WARN (and in Sentry breadcrumbs) will actually contain the row for
which the invalid accumulation was seen.