adapter: Route simple queries to mz_introspection (#20372)
Fixes #20247. Simple queries don't need a cluster to run on and it would
help some dbt use cases if we just handled them. The scope of "simple"
queries can increase over time, this PR handles the examples from the
ticket and unmaterializable functions which aren't evaluated on a
cluster anyway.