Add WASM-build support for JavaScript build targets to the lexer.
In an effort to ensure we're not maintaing two concurrent Lexer
implementations, make the existing lexer available to more platforms. By
building a WASM module, we can re-use it in JavaScript contexts.

`wasm-pack` does most of the heavy lifting, but we need to ensure we're
emitting types that can be serialized across the WASM interface, hence
the flattening of the token array.