Override from frankmcsherry/columnation to MaterializeInc/columnation (#20519)
See https://github.com/MaterializeInc/timely-dataflow/pull/2 for
context.

We want to keep the Materialize fork of timely buildable (which requires
making it depend on frankmcsherry/columnation, because columnation is
not available on crates.io), and we also want to avoid pointing it at
the Materialize fork of `columnation`. But we want to use the
Materialize fork of columnation in Materialize.