compute: tokenize mz_join_core
This commit adds a shutdown token check to the `mz_join_core` linear
join implementation. When the dataflow is shutting down, this makes the
operator discard all its existing work and new input data, rather than
processing it. As a result, differential join operators shut down faster
and emit less data, which in turn speeds up shutdown of downstream
operators.

Unfortunately, we can't make the same change for the DD join operator.
We could add a token check into the result closure we pass to that
operator, but the shutdown check would interfere with the fueling of the
DD join operator. Fuel is consumed based on the number of updates
emitted. When the token is dropped, the join closure stops producing
updates, which means the operator stops consuming fuel, so it does not
yield anymore until it has drained all its inputs. If there are many
inputs left, the replica may not accept commands for potentially quite a
long time.