Enable mz_join_core by default
Now that all cloud environments use mz_join_core, we should also flip
the default in the code, so that development and CI that doesn't fetch
LD flags tests the right implementation.

We still leave the flag around to allow us to switch back to the DD
implementation in a pinch.