Enable selected SQLite SLTs for auto_index_selects
This commit chooses two SQLite SLTs to be run in CI in the slow SLT
pipeline with --auto-index-selects. The two SLTs are selected based on
including min/max query patterns and being sufficiently small to
represent only a small execution time impact on the total test running
time. Thus, these tests can be used to increase the corpus used for
daily testing of monotonic one-shot SELECTs. Other tests in the SQLite
corpus that we run were considered. However, we observed that:
(a) There seem to be none with top-k patterns; (b) The other tests
with min/max patterns were about 10x larger in number of statements,
being a poor fit in terms of performance to be run without
--auto-transactions.