adapter: Check privileges before purifying
Privileges are usually checked after a statement has been planned.
When creating a source we may use a connection and/or secret as part of
the purification process. The purification process happens before
planning, which means we use the connects and secrets before checking
if we have the proper privileges.

This commit adds a privilege check specifically for CREATE SOURCE that
checks if the session has the necessary privileges to purify that
source. The check is done right before purification. Once the statement
is fully planned it will be checked again with a more comprehensive
list of privileges needed for creating a source. As a consequence there
will be some redundant privilege checks, but that's not a big deal
because it's fairly cheap to check.

Fixes #20606