sql: Implement pg_has_role function
This commit implements the pg_has_role function which reports if a role
has the privileges of another role through membership. In PostgreSQL
the function makes a distinction between INHERIT and NOINHERIT roles.
However, in Materialize we haven't implemented NOINHERIT, so we can't
make the distinction.

Documentation for this function can be found here:
https://www.postgresql.org/docs/15/functions-info.html

Resolves #20614