Prepared statement logging infrastructure
Sets up some infrastructure that will be necessary for statement
logging. Each prepared statement is associated with a `logging` object
that records whether it has been logged yet (this will always be
false, as of yet, since statement logging has not landed yet). If a
prepared statement has not been logged yet, this object records all
the information necessary to log it, as well as whether it has been
recorded yet in our statement logging prometheus metrics.

The reason why we keep around this information, rather than just
immediately logging a prepared statement as soon as it is prepared, is
because once statement logging exists, logging of prepared statements
will be controlled by logging of their _executions_. That is, when we
decide to log a statement execution (based on some sampling rate), we
need to log the corresponding prepared statement if and only if it has
not already been logged.

The unusual type of shared pointer (`Arc<QCell<T>>`) used to refer to
the logging metadata needs some comment.

First, we need some kind of reference-counted pointer, because
the lifetime of these objects is only knowable at runtime, and is not
associated with the lifetime of the prepared statement to which they
refer. For example, if the prepared statement is created, and then
bound to one or more portals, and then deallocated, we still need to
keep this logging metadata around, as one or more of the portals might
still be executed, requiring us to log the prepared statement (which
has now disappeared.)

Second, we cannot use `Rc<RefCell<T>>` for this pointer, because it
is sent to different threads (e.g., between the main coordinator task
and the pgwire tasks).

Third, it is inelegant/unnecessary to use something as heavyweight as
`Arc<Mutex<T>>`, because all references to the logging object will be
accessible from only one thread at any given time: the thread that has
control over the session object. `Arc<QCell<T>>` is a type that lets
us express these semantics precisely: the underlying `&mut T` is only
accessible via exclusive access to a `QCellOwner` object common to all
the references. This object is owned by the session, which gives us
the properties we need.