compute: mz_dataflow_initial_output_duration_seconds
This commit introduces a new replica metric,
`mz_dataflow_initial_output_duration_seconds`, that tracks the time from
the installation of a compute collection until it first produced any
outputs. This is not necessarily the time until a dataflow has been
fully hydrated (depending on your definition of 'hydration'), but might
be a good stand-in.