Allow more general Stream type in `RecordFirstRowStream`.
As part of the upcoming statement logging PR, I have different types
of row stream that I may want to return to the client (for example, if
the rows are immediately available). This PR unblocks that. As of now,
it does not change any behavior, but as it shouldn't cause any issues,
I hope folks won't mind landing it.