adapter: hook up group commit tracing span for table writes
Group commits can be kicked off by two things: a table write or the
advance_timelines_interval. For the former, hook up the tracing spans
such that the group commit gets the table write as a parent. For the
latter, give the group commit a new root span (with follows_from set
because why not I suppose).

This makes group commit (and the resulting persist operations) show up
in the trace given out by `emit_trace_id_notice` for the table write.

Also hook up a span for Command::Commit, so we get the commit traces
included for implicit transactions.