persist: disable crdb automatic stats on consensus table
These stats are for the cost-based optimizer but all the queries against
this table are single-table and very carefully tuned to hit the primary
index, so the cost-based optimizer doesn't really get us anything. OTOH,
the background jobs that crdb creates to collect these stats fill up the
jobs table (slowing down all sorts of things). So, disable them.