adapter: Prevent dangling refs in default priv
Default privileges contain optional references to databases and
schemas. Previously, when we dropped a database or schema, we didn't
check to see if there's a reference to that database or schema in the
default privileges tables. That allows default privileges to contain
dangling references to dropped databases and schemas.

This commit updates the DROP DATABASE and DROP SCHEMA command to revoke
default privileges if they contain references to the dropped database
or schema.

Fixes #20721