LIR refinement design doc for monotonic SELECTs
This commit revises the description related to LIR refinement of the
must_consolidate flag in the design document for stabilizing monotonic
one-shot SELECTs. In particular, it is observed that Plan::ArrangeBy
nodes are not monotonicity enforcers and that the situations in which
Plan::Get nodes are enforcers are also more restricted than previously
thought.