Fix: adapt the old mz to work with the new region api (#20687)
This PR updates the mz CLI to use the new Region API.

As a side effect, the concept of environment almost vanished from the code. Only remains the Environment  HTTP and SQL endpoints. The code does not yet implement the `cloud-api` nor the `frontegg-client`. The whole `mz` package will be replaced by another code. This is just a temporary patch.