Improvements to must_consolidate docstrings
This commit improves the docstrings for the must_consolidate flag by
noting that it is not a requirement that the operator implementations
perform consolidation of the input given to the operator prior to
internal transformations if the flag is set. The operator
implementations can perform any work that does not violate correctness
prior to consolidating. For example, Plan::Reduce can apply the
KeyValPlan and a streaming consolidate before honoring honoring the
must_consolidate setting.