adapter: Add usage checks to item schemas
This commit requires callers to have USAGE privileges on the schemas of
all connections, secrets, and types used in a query. Previously this
privilege was left out, but was often implicitly included if the target
of the query was in the same schema. For example,
`SELECT a::schema.ty FROM schema.table` would include the privilege
since `t` was in the same schema as `ty`.
`SELECT a::other_schema.ty FROM schema.table` would not include the
privilege because `t` was in a different schema as `ty`.