catalog: Fix custom array type names
For all custom types, we created two custom `postgres_types::Type`
types. One to represent the type itself and one to represent an array
containing that type. The naming convention of the array types were
`<type-name>_array`. However, in our catalog we stored the array types
with a naming convention of `_<type-name>`, which is how all non-custom
array types are named. If we ever tried to resolve an array of custom
types using the `<type-name>_array` format, it would fail.

This commit updates the naming convention of custom array types to
always be `_<type_name>`.

This never caused any issues previously because we never tried to
resolve a custom array using the name from `postgres_types::Type`. Had
we created a function that accepted a custom array type, then we would
have seen the issue.

Additionally, this commit updates the human-readable names of these
custom array types to match the human-readable names of non-custom
array types. That is, `<type-name>[]`.