explain: force fast path only in `(OPTIMIZED|PHYSICAL) PLAN`
Currently, our `EXPLAIN OPTIMIZER TRACE` logic will emit the fast path
plan for any stage that contains a `DataflowDescription<T>` instead of
the plan associated with that stage.

The purpose of this behavior is to emit fast path plans when people
request either `EXPLAIN FOR (OPTIMZIED|PHYSICAL) PLAN`. However, this
behavior is confusing when looking at  `EXPLAIN OPTIMZIER TRACE`
results.

This commit changes the behavior so now the `fast_path_plan` is emitted
only for `TraceEntry<T>` instances which correspond the `OPTIMIZED` or
`PHYSICAL` stages.

Fixes #19995.