Support building WASM crates and publishing them to npm.
Add CI hooks to build WASM packages alongside the crates in CI, and
tooling to publish built artifacts to npm with the appropriate tags and
version information.

This will publish packages for _all_ builds on the deploy pipeline.
Stable/official builds will not have a `-dev` suffix, and will receive
the `latest` dist tag. Unstable builds will have a `dev` suffix that
includes a build number, and will receive the `dev` dist tag.