adapter: Fix DROP OWNED with linked clusters
Previously, when trying to drop a source with a linked cluster via
`DROP OWNED`, `environmentd` would panic. The issue was that the
objects were being dropped in the wrong order. Linked clusters are
properly added to the list of dropped IDs when collecting the source's
dependencies, so we can skip over them when initially gathering objects
owned by the target role.

Fixes #20780