adapter: Keep linked cluster owners in sync
This commit ensures that the owner of a linked cluster, and all of its
replicas, are always the same as the owner of the linked object. This
commit prevents anyone from manually altering the owner of a linked
cluster or linked cluster replica. Additionally, whenever the owner of
a linked object is altered, the owners of the linked cluster and linked
cluster replicas are automatically altered.

Works towards resolving #20782