catalog: Fix stash item updates to persist entry
Previously, when updating an item in the stash, only the create SQL
definition would get updated. All other updates were ignored and
therefore lost after a restart.

The main issue was that the item update API only accepted the create
SQL statement and not the entire catalog item. This commit updates the
API so that the entire catalog item is considered.

Fixes #20795