adapter: Sync managed cluster and replica owners
This commit ensures that the owner of a managed cluster is always kept
in sync with the owner of its replicas. This commit prevents anyone
from manually altering the owner of a managed cluster replica.
Additionally, whenever the owner of a managed cluster is altered, the
owners of the replicas are automatically altered.

Fixes #20792