adapter: dump most catalog state
Use serde to dump nearly all of the CatalogState struct. This provides
much broader testdrive verification and also doesn't rely on humans to
remember to add serialization when dependent structs change.

One risk is that the serialization is not roundtrippable, and that's not
obvious to programmers who may assume they can add a `Deserialize` trait
to many of these objects and it'll work as expected.

Fixes #20811