cloud-api: handle successful request with no content (#20829)
The `cloud-api` client now has an additional logic to return an error
(`SuccesfullButNoContent`) and avoid parsing the JSON content for `204
(No Content)` requests.

The old approach tries to parse the content and produces a parsing
error, even when using an `Empty` struct with a custom `Deserialize`
implementation to "parse nothing". It seems that `No Content` != `Empty
Content`.

This change only affects the `GET` method for `/api/regions/`. The
`DELETE` method ha no problem accepting the `Empty` struct.