adapter: Keep replica owners in sync with cluster
This commit updates the semantics of cluster replica owners, so that a
replica owner is always kept in sync with the cluster owner. Alter the
owner of a replica is no longer permitted.

This fits into the model of managed clusters, where replicas are part
of the same logical object as the cluster.

Fixes #20792