storage-client: don't record dropped collections
The storage controller does not remove all collections from its state
when they are dropped. In particular, tables and source progress
collections remain in the state indefinitely.

We do not want to show frontiers for collections that have been dropped
in `mz_frontiers`, so we need to filter them out. Upon dropping a
collection, its read capability is set to the empty frontier. Thus this
is how we can check whether a given collection has been dropped or not.