sql: Require target role in alter default privs
Previously, the target role in an ALTER DEFAULT PRIVILEGES command was
optional, and would default to the current role when omitted. Most
users would get confused about the target role, omit it, and assume
the end result was what they wanted. Using the current role as the
target role was rarely what they intended.

This commit updates the syntax for ALTER DEFAULT PRIVILEGES so that the
target role is mandatory. This will require users to be explicit about
the target role and hopefully prevent them from incorrectly using the
current role.