Expose lexer keywords to WASM32 targets.
It is useful for external consumers to know what keywords Materialize
understands. For example, the SQL Shell can use this to apply accurate
syntax highlighting to SQL commands.

Unfortunately we cannot apply the wasm-bindgen proc macro directly to
the `KEYWORDS` const as the macro only works with functions, structs,
enums, impls, and extern blocks. As a result, this introduces a new
WASM-only function to expose the keywords to WASM32 consumers.

Touches MaterializeInc/console#304