Test single-time physical monotonicity enforcers
This commit adds an mzcompose cluster test that checks that no errors
are observed when non-consolidated data is given as input, then passed
to a single-time physical monotonicity enforcer, and then supplied to
a min/max reduction that will have must_consolidate turned off. The
latter is known since the queries used employ the same SQL patterns
that are checked with EXPLAIN PHYSICAL PLAN in
relax_must_consolidate.slt.