stash: Remove timestamps from public API
Internally the stash is implemented using timestamps for every key
value pair inserted/removed. Consumers of the stash treat the stash as
a key-value store and completely ignore the timestamps except for the
following scenarios:

- Checking if a stash collection is initialized by comparing the
collection's upper to `Timestamp::MIN`.
- Writing an empty batch to a stash collection at its upper to
initialize it.
- Dumping the raw contents of a stash collection for the stash-debug
tool.

While we may one day want to expose the Time Varying Collection (TVC)
internals of the stash, the current usage of timestamps by consumers,
other than the stash-debug tool, feels more like an abstraction leak
than a feature. Therefore, this commit updates consumers, other than
the stash-debug tool to not use timestamps, and remove the `pub`
visibility of the stash's timestamp internals.

The `TypedCollection::iter` method is the only method left that exposes
timestamps.