storage: Fix stash usage in finalize_shards
Shard finalization reads a list of shard IDs from the stash and
attempts to close all the shards returned.

Previously, the IDs were read from the stash using the
`Transaction::peek` method, which returns an unconsolidated list of
(key, value, diff) triples. The diffs were all ignored and the code
would attempt to close each key. As a result it was possible to attempt
to close the same shard twice and attempt to close an already closed
shard.

The shard finalization code looks like it correctly handles duplicate
and already closed shards. However, it's likely not the intent of
shard finalization to read unconsolidated stash data.

This commit updates shard finalization to use the
`Transaction::peek_one` method, which returns a consolidated list of
(key, value) pairs.