stash: Remove peek from public API
The peek method returns a consolidated list of (key, value, diff)
tuples from a stash collection. Previously, all usages of peek either
ignored the diff or asserted that it was equal to 1. The peek_one
method does exactly that, it returns an error if the consolidated diff
is not equal to 1 and only returns consolidated (key, value) pairs.

This commit updates all usages of peek with peek_one and removes the
`pub` modifier from peek.

In general all current consumers of the stash assume that the diff will
always be equal to 1. We could remove the concept of a diff from the
public API to simplify things, but it's not clear that it would be
beneficial and it would require a larger refactor of `AppendBatch`.