Drop the isolated runtime on its own thread
Sqllogictest shuts down the main tokio runtime, and expects that to also
stop all background tasks. However, Persist's own runtime does not shut
down synchronously, since it's against the rules to drop a runtime
within another runtime... which can cause background tasks to continue
to run while eg. filesystem state is deleted. This slightly horrible
commit forks off another thread just so we can drop the runtime
synchronously and outside the runtime context, but still allow the main
runtime to wait on it in a task (so that shutting down the outer runtime
will also block on the inner shutdown).