Update contributing guidelines to discourage contributions to certain areas (#20406)
As a follow up to
https://github.com/MaterializeInc/materialize/issues/19743, we want to
be more prescriptive about what types of external contributions we'll
encourage, and more explicit guidelines for picking up an issue.

Surfaces has decided to disallow external contributions to the
coordinator for two reasons:
* The cost of educating and performing extra-thorough reviews is not
worth the benefit-cost/risk tradeoff, given the complexity and
importance of this code
* Major coordinator changes should only be done as part of a more
focused, planned body of work, again given the nuances. One-off changes
picked up at random increase the risk of missed problems.

Questions for reviewer:
* Any other suggestions for both the "areas that are best suited for
external contributions" and "ask first" sections? I picked persist there
arbitrarily, and can remove if folks disagree.

This is one of a series of follow up items from the above issue.