adapter: opt pgwire queries and timeline advancement into INFO spans
We'd like to start defaulting to some small, select amount of always on
tracing. Things done directly on behalf of a user query (as well as the
periodic group commit run to advance timelines) seems like a nice place
to start getting a sense for what sort of volume our tracing infra will
support.