EXPECTED GROUP SIZE tuning test
This commit adds a testdrive test that documents a query over our
introspection sources to provide tuning advice for the EXPECTED GROUP
SIZE query hint. The query is validated by using the TPC-H load
generator, where there is uniformity in maximum group sizes for
different attributes in the lineitem table even under updates. As such,
while not true in general, it is possible for the tuning advice query
to hit tuning values that are exact in this case. The query observes
the hierarchies built for min/max/top-k patterns per dataflow and
corresponding region, and looks for how many levels do not contribute
to significant reductions in records maintained. Based on how many
levels can be cut and knowledge about fan-in (currently, 16), a query
hint value is computed per dataflow and relevant region. Note that by
following this procedure, the advice query can suggest that the
EXPECTED GROUP SIZE be tuned down, but not up. So it is important to
review the advice with some care when actually tuning, e.g., by
observing if update times are still acceptable and by taking into
account any domain knowledge about maximum group sizes that may
manifest in the future.