config: relax `synchronized` guards in `SystemParameterBackend`
This fixes a bug that was introduced by #20947, which excluded the
`enable_launchdarkly` from the set of synchronized parameters.

Because the `pull` methods in both `SystemParameterBackend` and
`SystemParameterFrontend` were guarding against modifications of
non-sycnhronized parameters, we were no longer able to pick up changes
to the `enable_launcharkly` flag in the `system_parameter_sync` loop.

To fix this issue, this commit relaxes the way `synchronized`-based
guardsd work as follows:

- The `SynchronizedParameters::modify` method does not apply the guard.
  Instead, it will modify any parameter, but only extend the `modified`
  set for (name, value) pairs where name is in the `synchronized` set.
- The `SystemParameterBackend::pull` no longer pulls entries only from
  the `synchronized` set. This allows us to pick up changes to the
  `enable_launchdarkly` flag even if it is marked as "not synchronized".
- Note that `SystemParameterFrontend::pull` will continue to pull and
  propagate changes from the LD client to the `SynchronizedParameters`
  instance **only if** they are contained in  the `synchronized` set.