adapter: fix `auto_run_on_introspection` for `EXPLAIN` plans
The `auto_run_on_introspection` method is missing a case for `EXPLAIN`
plans consistent with the `SELECT` handling. Consequently, the `EXPLAIN`
output for introspection queries does not consider indexes that are
actually picked up during the actual execution.

This commit adds this case. Note that due to the somewhat awkward
boundary between the planner and the coordinator (all `Plan` variants
except `Explain` wrap `Mir~` expressions), I had to re-define code that
is currently defined for `MirRelationExpr` terms to `HirRelationExpr`.

For faithful results, the two implementations should always be kept in
sync. I've added comments on the corresponding `Hir~` and `Mir~` code to
highlight this.