adapter,compute: unbatch compute commands
This commit removes the batching previously build into
`ComputeCommand`s. As a result, `CreateDataflow`, `AllowCompaction`, and
`CancelPeek` now take a single item, rather than a collection of items.
`CreateDataflow` and `CancelPeek` are also renamed from their plural
form.

This commit also adjusts the controller API accordingly, to simplify the
code a bit. This means that this change affects adapter code as well,
but only slightly.

The unbatching generally improves readability and reduces complexity by
removing loops over the batches. It also makes it easier to reason about
command counts in the `ComputeCommandHistory`, which previously required
thinking about "physical" and "logical" commands. Now these are the
same.