sql: Prevent users from executing dangerous funcs (#20998)
This commit prevents users from executing dangerous functions. For now,
a dangerous function is defined as any function found in the
mz_internal schema. In the future we could be more specific and add
attributes to each function to indicate if it's safe for a user to
evaluate.

Dangerous functions include functions that are only meant to be
executed in tests, such as mz_panic and mz_sleep, and functions
that are used to implement other functions/operators but don't
have well-defined semantics, such as mz_all and mz_any. If a user ever
executes one of these functions they would either get back
confusing/unhelpful results or crash part/all of their environment.

Fixes #15805
Fixes #18023
Fixes #18057

---------

Co-authored-by: nrainer-materialize <129728240+nrainer-materialize@users.noreply.github.com>