Thread time all the way through to BatchBuffer
This ensures any consolidated output is ordered by time.

One might imagine that since the KV ordering is based on the serialized
bytes, and therefore somewhat arbitrary, fine to do the same for T.
However, we often need to advance the timestamp to some frontier at read
time, and it is very useful if that is an order-preserving operation.