adapter: First step in renaming mz_introspection role to mz_system
I'm renaming the system role id that used to be called mz_introspection to mz_system.
This requires special logic in authentication to use the mz_system role when the mz_introspection
is provided but this logic can be removed as soon as no one uses mz_introspection.

Another approach I tried was to add a new system role, grant it the same access
that mz_introspection has and then eventually delete the mz_introspection role but couldn't
think of a way to avoid leaving a gap in the system ids assigned.