Remove enable_monotonic_oneshot_selects from tests
This commit eliminates the setting of the feature flag
enable_monotonic_oneshot_selects in all tests that used to reference
it. This is now a fixed feature for SLTs and enabled by default in all
mzcompose workflows. So the feature should always be exercised in both
CI and local SLT testing. This commit is a first step towards
eventually retiring the feature flag.