Remove enable_with_mutually_recursive from tests
This commit eliminates the setting of the feature flag
enable_with_mutually_recursive in all tests that used to reference it,
with exception of the SQL feature flag tests. This is now a fixed
feature for SLTs and enabled by default in all mzcompose workflows. So
the feature should always be exercised in both CI and local SLT
testing. This commit is a first step towards eventually retiring the
feature flag.