Introduce built-in mz_expected_group_size_advice
This commit creates a new built-in view to provide advice on tuning the
EXPECTED GROUP SIZE for dataflows running in a given replica that
include min/max/top-k maintenance patterns. The view comprises a large
query, with annotations, that exploits information from introspection
sources in determining if the levels of a hierarchical reduction can be
further cut down.