test: Deflake test_storage_usage_collection_interval_timestamps
The test_storage_usage_collection_interval_timestamps tests that after
starting up Materialize for the first time, an initial storage usage
collection is executed, and all rows generated use the same timestamp.

The storage usage interval used for the test used to be 5 seconds, so
if the test took longer than 5 seconds then we would have more than one
timestamp and mistakenly fail the test.

This commit updates the test so that if multiple timestamps are found,
we assert that they are taken at least storage usage interval
units apart.

Fixes #21074