storage: split persist write worker into table and monotonic
In particular, because the upcoming Pv2 txn management stuff wants to
operate on the set of all tables and this makes it _far_ easier to
reason about what's going on.

The intermediate state here is duplicative, but the impl of the table
worker will be completely replaced by #20954. Getting this refactor in
now so that no one breaks it in the meantime.

As far as I can tell, this really shouldn't have any behavior
difference:

- We hold a few less persist WriteHandles than we did before, but they
  were never used and heartbeats were recently removed from WriteHandles,
  so this amounts to an insignificant memory reduction.
- We no longer group up the monotonic and table appends into one big
  batch to concurrently run, instead grouping each into batches
  separately. From looking at traces, these seemed to end up getting
  batched separately (because of time) in practice, anyway.

We might also want to do the same sort of split for the read worker, but
I'm not sure yet, so leaving that out for now.