persist: fix "force keep" stats trimming bug
This was found by a customer who was trying to use pushdown. When it
didn't work, we looked at the stats and found the entire "ok" struct
containing the timestamp field had been trimmed, which is not how the
force_keep_col stuff is meant to work.

Concretely, trim_to_budget method previously only operates on the top
level struct columns. This (sorta) worked before #19309, but now there
are always two columns at the top level, "ok" and "err", and the real
columns are all nested under "ok". As a result, we'd end up trimming the
entire "ok" struct before even getting a chance to hand the timestamp
column to force_keep_col.

The fix is to make trim_to_budget recursive, also keeping any ancestors
of "force keep" columns.