adapter: Use timestamps in constant queries
Previously, the timestamp chosen by the Coordinator for constant
queries was always the maximum timestamp. The rationale was that
the answer to constant queries were known up to the end of time, so
we choose the freshest timestamp possible.

There was also some code gymnastics we did to all allow constant
queries in any type of transaction. Ignoring the timestamp for these
queries made this gymnastics easier.

Timestamps chosen for constant queries affect the answer if the query
has with a temporal filter or the query uses AS OF. So previously, we
would give the wrong answer for these types of queries.

This commit updates timestamp selection for constant queries so that we
pick a valid, potentially non-maximum, timestamp. The logic for
timestamp selection is the same as non-constant queries.

Fixes #18948