List replicas in SHOW CLUSTERS (#19104)
Add an array of replicas to the output of `SHOW CLUSTERS`. My reasoning
for wanting to add this is that it is the most commonly useful
information about a cluster. There has been some concern that this makes
our `SHOW` commands somewhat "denormalized", since we are now
duplicating information between `SHOW CLUSTERS` and `SHOW CLUSTER
REPLICAS`. However, my personal opinion is that it doesn't matter,
because the `SHOW` commands are meant to be broadly useful to users, not
to be as mathematically elegant as possible.

The rationale for which fields to include in the tuples are those that
are mandatory when creating a replica as part of the `CREATE CLUSTER`
command, i.e., name and size.

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered.
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
* Add replica sizes and names to `SHOW CLUSTERS`