adapter: Add metric for mz_introspection logins
When it goes to zero we can remove the hack that maps it to the mz_support user.
While implementing this I realized that you can't use the normal http interface
with mz_introspection/mz_system: you use the internal interface and always use mz_system.
So I removed support for logging in as mz_introspection