adapter: Use timestamps for temporal subscribes (#21101)
Previously, all subscribes that were timeline independent were
considered timestamps independent. This is not always true, if a
subscribe contains a temporal filter then it is timestamp dependent
even if it is timeline independent. This logic was already being done
for peeks but was left off of subscribes. This commit adds the logic to
promote timeline independent temporal subscribes to timestamps
dependent.

Fixes #18948