Non-persistent statement logging (#20862)
This PR causes a log of statements to be maintained but not persisted,
in the tables `mz_prepared_statement_history` and
`mz_statement_execution_history`. Logging is controlled by the session
variable `statement_logging_sample_rate`, constrained by the system
variable `statement_logging_max_sample_rate`.

It works by emitting a "begin" event when the statement begins
executing, and an "end" event when it ends. What it means for statement
execution to "end" is determined by the type of statement. For most
statements, it is the moment when control of the session is given back
to the protocol layer. However, for some other statements (notably
non-constant `SELECT`s, and `SUBSCRIBE`s), a more complex notion is
needed; for example, a `SELECT` is only considered finished after the
cluster has returned results, whereas a non-constant `SUBSCRIBE` is
never finished unless it errors or the client cancels it.

Thus, we need to pipe the statement logging ID through the various
commands and responses that implement these more complex protocols. This
is done by adding that ID to the `ExecuteContextExtra` struct. We also
make that struct log an error if it's dropped anywhere other than in the
coordinator code that handles finishing logging, to ensure that we don't
forget somewhere to end a statement.

A follow-up PR will persist the log in the stash, which we may later
decide to replace with storing it directly in Persist. In this PR the
log is in a built-in table, which is cleared on restart.